requirements.txt
streamlit==1.41.1
pydantic==2.10.4
promptlayer==1.0.34
python-dotenv==1.0.1
pydantic-settings==2.7.1
openai==1.59.3
